eews_backend/requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiokafka==0.8.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
colorlog==6.7.0
confluent-kafka==2.2.0
contourpy==1.1.0
croniter==1.4.1
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.100.0
faust-streaming==0.10.14
fonttools==4.41.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
haversine==2.8.0
httptools==0.6.0
idna==3.4
importlib-resources==6.0.0
influxdb-client==1.36.1
intervaltree==3.1.0
Jinja2==3.1.2
kafka-python==2.0.2
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mode-streaming==0.3.5
motor==3.2.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
obspy==1.4.0
opentracing==2.4.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pycodestyle==2.10.0
pydantic==2.0.2
pydantic_core==2.1.2
pyflakes==3.0.1
pymongo==4.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
reactivex==4.0.4
requests==2.31.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.18
starlette==0.27.0
terminaltables==3.1.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
venusian==3.0.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.16.1
